darkmode settheme and theme moved to app.jsx and data passed down to nav and darkmode.jsx
