+++ client/src/components/DarkMode.jsx
@@ -1,29 +1,11 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-import { createTheme, ThemeProvider } from '@material-ui/core/styles';
 import Switch from '@mui/material/Switch';
-import Typography from '@material-ui/core/Typography';
 import Paper from '@material-ui/core/Paper';
 import LightModeIcon from '@mui/icons-material/LightMode';
 import ModeNightIcon from '@mui/icons-material/ModeNight';
 
-const DarkMode = ({authUser}) => {
-
-  const [darkMode, setDarkMode] = useState(null);
-
-  const theme = createTheme({
-    palette: {
-      type: darkMode ? 'dark' : 'light',
-    },
-  });
-
-  const setTheme = ( themeBol ) => {
-    if (!themeBol) {
-      setDarkMode(false);
-    } else {
-      setDarkMode(true);
-    }
-  };
+const DarkMode = ({authUser, darkMode, setTheme, theme }) => {
 
   const addTheme = (darkMode) => {
     axios.put(`/users/theme/${authUser}`, {theme: darkMode})
@@ -37,13 +19,11 @@ const DarkMode = ({authUser}) => {
   };
 
   return (
-    <ThemeProvider theme={theme}>
-      <Paper elevation={0}>
-        <LightModeIcon style={{ color: darkMode ? 'grey' : '#f4c732' }}></LightModeIcon>
-        <Switch onChange={handleToggle} />
-        <ModeNightIcon style={{ color: darkMode ? '#40457e' : 'grey' }}></ModeNightIcon>
-      </Paper>
-    </ThemeProvider>
+    <Paper elevation={0}>
+      <LightModeIcon style={{ color: darkMode ? 'grey' : '#f4c732' }}></LightModeIcon>
+      <Switch onChange={handleToggle} />
+      <ModeNightIcon style={{ color: darkMode ? '#40457e' : 'grey' }}></ModeNightIcon>
+    </Paper>
   );
 
 };
